tests: add tests for integer parsing

+++ crates/mips_parser/src/lexer/defs.rs
@@ -24,5 +24,5 @@ pub enum Token {
     Register(Register),
     Ident(String),
     String(String),
-    Number(u64),
+    Number(i16),
 }
